add cookie.rb file with empty class definition
It's a best practice to have a single class in a given ruby file that has the same name of the class after converting the name to a snake case naming convention.

In our example here we name the file `cookie.rb` as our class name is `Cookie`.

Class names in Ruby follow a camel case naming convention. The first letter of the Ruby class must start with a capital letter as classes are considering "constants" in Ruby.