Add a public instance method
We simply define the method inside the class (before the `end`) in order to have a public instance method for our class.

We call it an `instance` method because this method applies on object created by the class, for instance:
```ruby
c = Cookie.new
c.eat
```
We call it a public method because I can call it in files and methods outside this class, for instance I can ruby the code above from IRB or from another class's method.