Add a private instance method
To add a private method write the keyword `private` somewhere within the class block. Any method that is defined between the `private` keyword and the `end` keyword of the class is treated as a private method.

Private methods are not supposed to be called from outside the class. You will get an error trying to access a private method directly:
```shell
NoMethodError: private method `bake' called for #<Cookie:0x0000010411b9a8>
```
Ruby gives programmer a lot of flexibilty. For instance you're able to call a private method using the `send` method. However, you should try to avoid that unless you have a strong reason to.

The main use of private methods is to follow best practices of object oriented design principles. This is a big topic, for now keep in mind this rule of thumb:
* Keep methods as private as possible