Call the private method from a public method
You're able to call a private method from a public method with in the same class. This way you can use a private method from outside the class without directly calling it.

In this example `bake_and_eat` calls the `bake` and the `eat` methods. This works fine because `bake_and_eat` method is defines within the same class.