Add attribute methods
It's generally useful to have the ability to read and write the instance variables defined in class from the outside. We can things of those variable as `attributes` of the object. We can simple read and write them using public methods.

However, there is a nicer shortcut in Ruby and that is to use `attr_reader` method which defines a method that returns the value of an instance variable with same name.

We can also define an `attr_writer` which defines a method that gives us the abiltiy to set an instance variable with the same name.

We can also define an `attr_accessor` which is the same as defining both an `attr_reader` and an `attr_writer`.